refactor(frontend): разнос lib/types.ts по доменным файлам — чистые перемещения
Этап F1 реструктуризации frontend: types.ts (2345 строк, 205 деклараций)
разнесён по region-границам в lib/types/*: tunnels, routing, servers,
system, deviceproxy, diagnostics, singbox, sbRouter, subscriptions.

lib/types.ts стал баррелем (export * from './types/...') — путь импорта
$lib/types сохранён, все 293 импортёра не тронуты. Cross-file ссылки
(UsageLevel в system.ts) — явными type-импортами.

Эквивалентность верифицирована мультимножествами строк кода и
комментариев: расхождения — только перенесённый import UsageLevel и
декоративные разделители. svelte-check 0/0, eslint 0, vitest 1343,
production build зелёный.
